implement get current weather service method

+++ dto/src/main/kotlin/org/stacklabs/weather/dto/CurrentWeatherDto.kt
@@ -15,5 +15,5 @@ data class CurrentWeatherDto(
     val windSpeed: Float,
     @PositiveOrZero
     @Schema(description = "Humidity in %")
-    val humidity: Float
+    val humidity: Int
 )
